backend/requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
anthropic==0.40.0
httpx==0.27.2
pydantic==2.10.2
databricks-sql-connector==3.4.0
python-dotenv==1.0.1
yfinance==0.2.54
